Fix example code in NuGet package
Change card number for a valid Stripe test card number.
Add CVC code to charge creation (as this is a required field for UK
Stripe accounts)
Change amount as although Stripe API is in cents, internally, amounts
are expressed in dollars.
Fix error handling as we need to check that IsError is false before we
can safely access the Paid boolean property.